api_django/requirements.txt
asgiref==3.7.2
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
openpyxl==3.1.2
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pypng==0.20220715.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0

# Cache Redis (opcional)
django-redis==5.4.0
